Refactor common code for jenkins build scripts
