+++ .buildbot/Jenkinsfile
@@ -46,10 +46,13 @@ def call() {
       /* Docker related env vars */
       DOCKER_PUSHPULL_RWLOCK = "/var/lock/jenkins-pipeline-${GITHUB_REPO_NAME}-${GITHUB_PR_BASEREF}.lock"
       DOCKER_DAEMON_SOCKET = 'unix://var/run/docker.sock'
-      DOCKER_REGISTRY_TOKEN_ACCESS = 'true'
-      DOCKER_REGISTRY_HOST_NAME = 'ghcr.io'  /* set to '' to use docker.io */
+      DOCKER_REGISTRY_HOST_NAME = 'ghcr.io'  /* settings for ghcr.io */
       DOCKER_REGISTRY_USER_NAME = 'onnxmlir'
       DOCKER_REGISTRY_LOGIN_NAME = 'onnxmlir'
+      # DOCKER_REGISTRY_HOST_NAME = ''  /* settings for docker.io */
+      # DOCKER_REGISTRY_USER_NAME = 'onnxmlirczar'
+      # DOCKER_REGISTRY_LOGIN_NAME = 'onnxmlirczar'
+      DOCKER_REGISTRY_TOKEN_ACCESS = 'true'
 
       /* Credentials defined in Jenkins */
       DOCKER_REGISTRY_LOGIN_TOKEN = credentials('GITHUB-ONNXMLIR-TOKEN')
@@ -98,7 +101,7 @@ def call() {
       JENKINS_RWLOCK="${JENKINS_SCRIPT_DIR}/jenkins-rwlock.py"
       JENKINS_STOP_PREVIOUS_BUILD="${JENKINS_SCRIPT_DIR}/jenkins-stop-previous-build.py"
       JENKINS_BUILD_LLVM_PROJECT="${JENKINS_SCRIPT_DIR}/jenkins-build-llvm-project.py"
-      JENKINS_BUILD_PROJECT="${JENKINS_SCRIPT_DIR}/jenkins-build-project.py"
+      JENKINS_BUILD_ONNX_MLIR="${JENKINS_SCRIPT_DIR}/jenkins-build-onnx-mlir.py"
       JENKINS_VERIFY_USER_IMAGE="${JENKINS_SCRIPT_DIR}/jenkins-verify-user-image.py"
       JENKINS_PUBLISH_DOCKER_IMAGES="${JENKINS_SCRIPT_DIR}/jenkins-publish-docker-images.py"
       JENKINS_CHECK_MODEL_ZOO="${JENKINS_SCRIPT_DIR}/jenkins-check-model-zoo.py"
@@ -197,11 +200,11 @@ def call() {
         }
       }
 
-      stage('Build project images') {
+      stage('Build onnx-mlir images') {
         when { not { environment name: 'GITHUB_PR_ACTION',
                      value: 'closed' } }
         steps {
-          sh "${JENKINS_BUILD_PROJECT}"
+          sh "${JENKINS_BUILD_ONNX_MLIR}"
         }
       }
