+++ .buildbot/jenkins-check-model-zoo.py
@@ -1,32 +1,7 @@
 #!/usr/bin/env python3
 
-import datetime
-import git
-import logging
-import math
-import os
-import requests
-import subprocess
-import sys
+from jenkins_common import *
 
-logging.basicConfig(
-    level=logging.INFO,
-    format="[%(asctime)s][%(lineno)03d] %(levelname)s: %(message)s",
-    datefmt="%Y-%m-%d %H:%M:%S",
-)
-
-# Set parallel jobs based on both CPU count and memory size.
-# Because using CPU count alone can result in out of memory
-# and get Jenkins killed. For example, we may have 64 CPUs
-# (128 threads) and only 32GB memory. So spawning off 128
-# cc/c++ processes is going to quickly exhaust the memory.
-#
-# Algorithm: NPROC = min(2, # of CPUs) if memory < 8GB, otherwise
-#            NPROC = min(memory / 8, # of CPUs)
-MEMORY_IN_GB = os.sysconf("SC_PAGE_SIZE") * os.sysconf("SC_PHYS_PAGES") / (1024.0**3)
-NPROC = str(math.ceil(min(max(2, MEMORY_IN_GB / 8), os.cpu_count())))
-
-DOCKER_DEV_IMAGE_WORKDIR = "/workdir"
 ONNX_MLIR_SOURCE = "/workdir/onnx-mlir"
 ONNX_MLIR_HOME = "/workdir/onnx-mlir/build/Debug"
 RUN_ONNX_MODEL_PY = "RunONNXModel.py"
@@ -35,58 +10,36 @@ RUN_ONNX_MODELZOO_PY = "RunONNXModelZoo.py"
 RENDERJSON_URL = "https://raw.githubusercontent.com/caldwell/renderjson/master/"
 RENDERJSON_JS = "renderjson.js"
 
-docker_daemon_socket = os.getenv("DOCKER_DAEMON_SOCKET")
-docker_registry_host_name = os.getenv("DOCKER_REGISTRY_HOST_NAME")
-docker_registry_user_name = os.getenv("DOCKER_REGISTRY_USER_NAME")
-github_repo_name = os.getenv("GITHUB_REPO_NAME")
-github_repo_name2 = os.getenv("GITHUB_REPO_NAME").replace("-", "_")
-github_pr_baseref = os.getenv("GITHUB_PR_BASEREF")
-github_pr_baseref2 = os.getenv("GITHUB_PR_BASEREF").lower()
-github_pr_number = os.getenv("GITHUB_PR_NUMBER")
-jenkins_home = os.getenv("JENKINS_HOME")
-job_name = os.getenv("JOB_NAME")
-workspace_dir = os.getenv("WORKSPACE")
-
 modelzoo_reportdir = os.getenv("MODELZOO_REPORTDIR")
 modelzoo_workdir = os.getenv("MODELZOO_WORKDIR")
 modelzoo_html = os.getenv("MODELZOO_HTML")
 modelzoo_stdout = os.getenv("MODELZOO_STDOUT")
 
-docker_dev_image_name = (
-    github_repo_name
-    + "-dev"
-    + ("." + github_pr_baseref2 if github_pr_baseref != "main" else "")
-)
-docker_dev_image_tag = github_pr_number.lower()
 docker_dev_image_full = (
     (docker_registry_host_name + "/" if docker_registry_host_name else "")
     + (docker_registry_user_name + "/" if docker_registry_user_name else "")
     + docker_dev_image_name
     + ":"
-    + docker_dev_image_tag
+    + pr_image_tag
 )
 
 # History directory is just the job directory
-workspace_historydir = os.path.join(jenkins_home, "jobs", job_name)
-container_historydir = os.path.join(DOCKER_DEV_IMAGE_WORKDIR, job_name)
+workspace_historydir = os.path.join(jenkins_home, "jobs", jenkins_job_name)
+container_historydir = os.path.join(DOCKER_DEV_IMAGE_WORKDIR, jenkins_job_name)
 
-workspace_reportdir = os.path.join(workspace_dir, modelzoo_reportdir)
+workspace_reportdir = os.path.join(jenkins_workspace_dir, modelzoo_reportdir)
 container_reportdir = os.path.join(DOCKER_DEV_IMAGE_WORKDIR, modelzoo_reportdir)
 
-workspace_workdir = os.path.join(workspace_dir, modelzoo_workdir)
+workspace_workdir = os.path.join(jenkins_workspace_dir, modelzoo_workdir)
 container_workdir = os.path.join(DOCKER_DEV_IMAGE_WORKDIR, modelzoo_workdir)
 
-workspace_model_py = os.path.join(workspace_dir, "utils", RUN_ONNX_MODEL_PY)
+workspace_model_py = os.path.join(jenkins_workspace_dir, "utils", RUN_ONNX_MODEL_PY)
 container_model_py = os.path.join(DOCKER_DEV_IMAGE_WORKDIR, RUN_ONNX_MODEL_PY)
 
-workspace_modelzoo_py = os.path.join(workspace_dir, "utils", RUN_ONNX_MODELZOO_PY)
+workspace_modelzoo_py = os.path.join(
+    jenkins_workspace_dir, "utils", RUN_ONNX_MODELZOO_PY
+)
 container_modelzoo_py = os.path.join(DOCKER_DEV_IMAGE_WORKDIR, RUN_ONNX_MODELZOO_PY)
-
-
-def urlretrieve(remote_url, local_file):
-    req = requests.get(remote_url)
-    with open(local_file, "wb") as f:
-        f.write(req.content)
 
 
 def main():
